Replace custom service registry with ServiceLoader pattern
Discover PurchasesService implementations via java.util.ServiceLoader
instead of the imperative Purchases.registerService/unregisterService
APIs and the custom PurchasesServiceRegistry.

- Core: PurchasesServices default forwarder now loads providers with
  ServiceLoader.load(PurchasesService::class.java, ...classLoader);
  removed PurchasesServiceRegistry and the (internal) register/
  unregisterService methods on both Purchases variants.
- AdMob: added RewardVerificationService provider declared in
  META-INF/services plus a consumer ProGuard keep rule; dropped the
  init-time registerService call.

Resolves ADS-277.
