+++ feature/admob/src/main/resources/META-INF/services/com.revenuecat.purchases.PurchasesService
@@ -1,0 +1,1 @@
+com.revenuecat.purchases.admob.rewardverification.RewardVerificationService
