+++ purchases/src/test/java/com/revenuecat/purchases/PurchasesLifecycleWiringTest.kt
@@ -9,7 +9,6 @@ import org.junit.Test
 @OptIn(InternalRevenueCatAPI::class)
 internal class PurchasesLifecycleWiringTest {
     private lateinit var originalServiceForwarder: PurchasesService
-    private val registeredServices = mutableListOf<PurchasesService>()
 
     @Before
     fun setUp() {
@@ -18,8 +17,6 @@ internal class PurchasesLifecycleWiringTest {
 
     @After
     fun tearDownMocks() {
-        registeredServices.forEach { Purchases.unregisterService(it) }
-        registeredServices.clear()
         Purchases.serviceForwarder = originalServiceForwarder
         Purchases.backingFieldSharedInstance = null
     }
@@ -37,40 +34,6 @@ internal class PurchasesLifecycleWiringTest {
         verify(exactly = 1) {
             serviceForwarder.close(purchases)
             orchestrator.close()
-        }
-    }
-
-    @Test
-    fun `registerService receives lifecycle events from the default service forwarder`() {
-        val service = mockk<PurchasesService>(relaxed = true)
-        val purchases = mockk<Purchases>(relaxed = true)
-
-        Purchases.registerService(service)
-        registeredServices += service
-
-        originalServiceForwarder.initialize(purchases)
-        originalServiceForwarder.close(purchases)
-
-        verify(exactly = 1) {
-            service.initialize(purchases)
-            service.close(purchases)
-        }
-    }
-
-    @Test
-    fun `unregisterService stops further lifecycle delivery`() {
-        val service = mockk<PurchasesService>(relaxed = true)
-        val purchases = mockk<Purchases>(relaxed = true)
-
-        Purchases.registerService(service)
-        Purchases.unregisterService(service)
-
-        originalServiceForwarder.initialize(purchases)
-        originalServiceForwarder.close(purchases)
-
-        verify(exactly = 0) {
-            service.initialize(purchases)
-            service.close(purchases)
         }
     }
 }
